requirements-local.txt
# Local development requirements for face recognition setup
# Install with: pip install -r requirements-local.txt

# Core dependencies
numpy>=1.26.0
pandas>=2.2.0
pyyaml>=6.0
scikit-learn>=1.4.0

# Image processing (choose one or both)
opencv-python-headless>=4.9.0  # For advanced face detection
Pillow>=10.0.0  # For basic image processing

# Kaggle API for dataset download
kaggle>=1.5.0

# Optional: for better performance
matplotlib>=3.7.0
seaborn>=0.12.0
